requirements.txt
# --- Core Framework ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# --- Database (Async) ---
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1  # For Database Migrations

# --- Configuration & Env ---
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1

# --- Security (JWT & Password Hashing) ---
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9  # Required for Login Form Data

# --- Utilities ---
email-validator>=2.1.0